JS files support the statement @import
Even if the @import statement is not a javascript statement, it is
useful to have this feature in order to combine multiple .js files in
one file.
The new statement has exactly the same syntax as for CSS files (but
without support for media queries).
Example: @import url("filename.js");