Added more css files to the test website in order to reproduce bugs
2 bugs:
- Absolute relative path: without the fix I have submitted some minutes
ago, the second div will not display the background image (file
css-style5.css) because this file is imported from a file that it is in
the same directory.
- Cache dependencies don't work properly: To reproduce the error, load
the test web page, after modify the file css-style3.css, after modify
the file css-style4.css, and again the file css-style3.css and you will
see that cruncher is not detecting the change and doesn't create a new
version. As far as I had time to investigate, it seems there is a bug in
the file cache dependencies (file monitors).