Optimisation regarding the new "Physical files" feature
We have detected that with the new feature, an I/O operation is going to
be performed for every request. For a very busy site could be be
thousands of requests per minute, which means thousands of I/O
operations.
This optimization reduces the I/O operations to 1 per minute (for any
resource that has been previously cached and crunched)